units ai

+++ Assets/_Project/Scripts/InGame/GUI/ShowDamageText.cs
@@ -4,7 +4,7 @@ public class ShowDamageText : MonoBehaviour
 {
     public DynamicTextData TextData;
 
-    public void ShowDamage(int dmg)
+    public void ShowDamage(float dmg)
     {
         DynamicTextManager.CreateText2D(transform.position, dmg.ToString(), TextData);
     }
